+++ metalearners/xlearner.py
@@ -85,7 +85,7 @@ class XLearner(_ConditionalAverageOutcomeMetaLearner):
         n_jobs_base_learners: int | None = None,
     ) -> Self:
         self._validate_treatment(w)
-        self._validate_outcome(y)
+        self._validate_outcome(y, w)
 
         self._treatment_variants_indices = []