+++ CHANGELOG.rst
@@ -18,6 +18,10 @@ Changelog
 
 * Changed the index columns order in ``MetaLearnerGridSearch.results_``.
 
+* Raise a custom error if only one class is present in a classification outcome.
+
+* Raise a custom error if there are some treatment variants which have seen classification outcomes which have not appeared for some other treatment variant.
+
 0.6.0 (2024-07-08)
 ------------------